Upgrade react-qr-reader
