+++ package.json
@@ -5,7 +5,8 @@
   "main": "javascript/bundle.js",
   "dependencies": {
     "react": "^15.0.2",
-    "react-dom": "^15.0.2"
+    "react-dom": "^15.0.2",
+    "react-qr-reader": "^0.3.1"
   },
   "devDependencies": {
     "babel-loader": "^6.2.4",
